requirements.txt
Flask
requests
Pillow
tkinter
python-dotenv